requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
boolean.py==5.0
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
license-expression==30.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
openai==1.76.0
packageurl-python==0.16.0
packaging==25.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
platformdirs==4.3.7
pluggy==1.5.0
py-serializable==2.0.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-order==1.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
